common: Move FLAC metadata readers to metadata crate.
